GRILL 137 Fix SessionManager.stop()

+++ service/src/java/org/apache/hive/service/cli/session/SessionManager.java
@@ -114,7 +114,7 @@ public class SessionManager extends CompositeService {
     super.stop();
     if (backgroundOperationPool != null) {
       backgroundOperationPool.shutdown();
-      int timeout = hiveConf.getIntVar(ConfVars.HIVE_SERVER2_ASYNC_EXEC_SHUTDOWN_TIMEOUT);
+      long timeout = hiveConf.getLongVar(ConfVars.HIVE_SERVER2_ASYNC_EXEC_SHUTDOWN_TIMEOUT);
       try {
         backgroundOperationPool.awaitTermination(timeout, TimeUnit.SECONDS);
       } catch (InterruptedException e) {
@@ -315,3 +315,4 @@ public class SessionManager extends CompositeService {
     }
   }
 }
+
